pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "vntdr"
version = "0.1.0"
description = "Research-first quantitative trading foundation built around VeighNa."
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "httpx>=0.27,<0.29",
  "pydantic>=2.10,<3.0",
  "pydantic-settings>=2.7,<3.0",
  "python-okx>=0.4.1,<0.5",
  "python-dateutil>=2.9,<3.0",
  "redis>=5.2,<6.0",
  "sqlalchemy>=2.0.38,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "tenacity>=9.0,<10.0",
  "typer>=0.15,<1.0",
]

[project.optional-dependencies]
veighna = [
  "vnpy==4.3.0",
  "vnpy_ctastrategy==1.4.1",
  "vnpy_okx==2026.2.28",
  "vnpy_postgresql==1.1.2",
  "vnpy_riskmanager==2.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-cov>=6.0,<7.0",
]

[project.scripts]
vntdr = "vntdr.cli:run"

[tool.hatch.build.targets.wheel]
packages = ["src/vntdr"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
